refactor(suite): fee component
