+++ packages/suite/src/components/suite/modals/ReduxModal/UserContextModal/TxDetailModal/CancelTransaction/CancelTransaction.tsx
@@ -68,7 +68,11 @@ export const CancelTransaction = ({ tx, selectedAccount }: CancelTransactionProp
                         <Row gap={spacings.md}>
                             <Translation id="TR_CANCEL_TX_FEE" />
                             <Text variant="tertiary">
-                                <FeeRate feeRate={feePerByte} networkType={networkType} />
+                                <FeeRate
+                                    feeRate={feePerByte}
+                                    networkType={networkType}
+                                    symbol={tx.symbol}
+                                />
                             </Text>
                         </Row>
                     }
