+++ packages/suite/src/components/wallet/Fees/index.tsx
@@ -1,0 +1,23 @@
+import { CurrentFee } from './CustomFee/CurrentFee';
+import { CustomFee } from './CustomFee/CustomFee';
+import { CustomFeeEthereum } from './CustomFee/CustomFeeEthereum';
+import { CustomFeeMisc } from './CustomFee/CustomFeeMisc';
+import { CustomFeeWrapper } from './CustomFee/CustomFeeWrapper';
+import { BitcoinFeeCards } from './StandardFee/BitcoinFeeCards';
+import { EthereumFeeCards } from './StandardFee/EthereumFeeCards';
+import { FeeCard } from './StandardFee/FeeCard';
+import { MiscFeeCards } from './StandardFee/MiscFeeCards';
+import { StandardFee } from './StandardFee/StandardFee';
+
+export {
+    CustomFee,
+    CurrentFee,
+    CustomFeeEthereum,
+    CustomFeeMisc,
+    CustomFeeWrapper,
+    BitcoinFeeCards,
+    EthereumFeeCards,
+    FeeCard,
+    MiscFeeCards,
+    StandardFee,
+};
